+++ src/App.js
@@ -2,12 +2,26 @@ import Header from './components/Header';
 import Routes from './Routes';
 import GlobalStyles from './styles/globalStyles';
 
+import Sidebar from './components/Sidebar';
+import {
+  ContentWrapper,
+  LayoutContainer,
+  SideBarWrapper,
+} from './styles/layouts';
+
 function App() {
   return (
     <>
       <GlobalStyles />
       <Header />
-      <Routes />
+      <LayoutContainer>
+        <SideBarWrapper>
+          <Sidebar />
+        </SideBarWrapper>
+        <ContentWrapper>
+          <Routes />
+        </ContentWrapper>
+      </LayoutContainer>
     </>
   );
 }
