+++ src/components/SideBarTodoBox/styled.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const TodoContentWrapper = styled.div`
+  box-shadow: 0 3px 6px 0 rgba(0, 0, 0, 0.16);
+  background-color: #fff;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  border-radius: 30px;
+  padding: 20px 30px;
+  min-height: 300px;
+`;
+export const ListTitle = styled.p`
+  font-size: 1rem;
+  color: #424242;
+  display: flex;
+`;
