+++ src/components/Todo/TodoInsult.js
@@ -1,53 +1,58 @@
-import React, { Component, useCallback, useState } from 'react';
-import SelectBox from './SelectBox';
-import Modal from 'react-modal';
-import Calendar from 'react-calendar';
+import React, { useCallback, useState } from 'react';
 import {
   AddBtn,
   CalendarBtn,
-  CateBtn,
   Hr,
   MemoInput,
-  SelectBoxs,
   TagBtn,
   TodoInput,
   TodoInputWrapper,
 } from './styled';
-// SelectBox에 props로 보낼 옵션값들
-const selectOptions = ['학교', '동아리', '취미'];
-// 값을 받기 전, 값이 전부 공백인 객체형태의 state
-const TodoInsult = ({ onInsert }) => {
+import { todoState } from '../../store/todoState';
+
+const TodoInsult = () => {
   const [inputs, setInputs] = useState({
     content: '',
     memo: '',
-    cateAdd: '',
   });
-  const { content, cateAdd } = inputs; // 비구조화 할당을 통해 값 추출
-  // 키보드 입력을 받는 함수
+  const { addTodo } = todoState();
+
+  const { content, memo } = inputs; // 비구조화 할당을 통해 값 추출
+
   const onChange = (e) => {
-    const { value, name } = e.target; //e.target에서 name,value 추출
+    const { value, name } = e.target; // 우선 e.target 에서 name 과 value 를 추출
     setInputs({
-      ...inputs, //기존의 input 객체 복사
-      [name]: value, //name 키를 가진 값을 value 로 설정
+      ...inputs, // 기존의 input 객체를 복사한 뒤
+      [name]: value, // name 키를 가진 값을 value 로 설정
     });
   };
-  //모달 세팅
-  const [modalIsOpen, setModalIsOpen] = useState(false);
-  const [modalIsOpens, setModalIsOpens] = useState(false);
-  // todo 제출 함수
-  const onSubmit = useCallback(
-    (e) => {
+
+  const addTodoHandler = (e) => {
+    //내용이 입력되었으면 TODO 추가
+    if (inputs.content) {
+      addTodo(inputs);
       setInputs({
         content: '',
         memo: '',
-        cateAdd: '',
       });
-      onInsert(inputs);
+    } else {
+      //내용이 없으면 알림
+      alert('일정은 필수적으로 입력해주세요.');
+    }
+    e.preventDefault();
+  };
+  const onSubmit = useCallback(
+    (e) => {
+      addTodo(inputs);
+      setInputs({
+        content: '',
+        memo: '',
+      });
       e.preventDefault();
-      selectOptions.concat(cateAdd);
     },
-    [onInsert, inputs],
+    [addTodo, inputs],
   );
+
   return (
     <TodoInputWrapper>
       <form>
@@ -60,55 +65,22 @@ const TodoInsult = ({ onInsert }) => {
           onChange={onChange}
         />
         {/* 추가 버튼 */}
-        <AddBtn onClick={onSubmit} />
+        <AddBtn onClick={addTodoHandler} />
         <Hr />
         {/* 메모 작성 */}
-        {/*카테고리 부분 SelectBox*/}
-        <SelectBox
-          // 카테고리값 설정
-          value={inputs.memo}
-          setValue={(value) => {
-            setInputs((state) => ({
-              ...state,
-              memo: value,
-            }));
-          }}
-          // 추가로 셀렉박스에 넘겨줄 리스트를 props에 넣어준다.
-          options={selectOptions}
-        />
-        {/*
         <MemoInput
-          placeholder="카테고리추가(test)"
-          name="cateAdd"
+          placeholder="상세설명"
+          name="memo"
           type={'text'}
-          value={cateAdd}
+          value={memo}
           onChange={onChange}
           size={'0.8rem'}
           height={'2rem'}
         />
-        */}
-        {/* 카테고리 버튼 <TagBtn />*/}
-        {/* 달력 버튼 <CalendarBtn />*/}
-        <>
-          <CateBtn onClick={() => setModalIsOpens(true)}>카테고리 수정</CateBtn>
-          <Modal isOpen={modalIsOpens}>
-            <h1>추가할 카테고리명</h1>
-            <input />
-            <button onClick={() => setModalIsOpens(false)}>추가</button>
-            <hr></hr>
-            <h1>삭제할 카테고리명</h1>
-            <input />
-            <button onClick={() => setModalIsOpens(false)}>삭제</button>
-          </Modal>
-        </>
-        <>
-          <CateBtn onClick={() => setModalIsOpen(true)}>달력</CateBtn>
-          <Modal isOpen={modalIsOpen}>
-            <Calendar></Calendar>
-            <hr />
-            <button onClick={() => setModalIsOpen(false)}>날짜 입력</button>
-          </Modal>
-        </>
+        {/* 카테고리 버튼 */}
+        <TagBtn />
+        {/* 달력 버튼 */}
+        <CalendarBtn />
       </form>
     </TodoInputWrapper>
   );
