+++ src/components/Todo/TodoList.js
@@ -1,8 +1,10 @@
 import React from 'react';
 import TodoListItem from './TodoListItem';
 import { TodoListWrapper } from './styled';
+import { todoState } from '../../store/todoState';
 
-const TodoList = ({ todos, onRemove, onToggle }) => {
+const TodoList = ({ onRemove, onToggle }) => {
+  const todos = todoState((state) => state.todos);
   return (
     <TodoListWrapper>
       {todos.map((todo) => (
